moved WatchlistGQL to queries
