+++ src/ducks/Watchlists/GetFeaturedWatchlists.js
@@ -1,5 +1,5 @@
 import getEnhancedWatchlist from './getEnhancedWatchlist'
-import { FEATURED_WATCHLIST_QUERY } from './../../components/WatchlistPopup/WatchlistGQL'
+import { FEATURED_WATCHLIST_QUERY } from '../../queries/WatchlistGQL'
 
 export default getEnhancedWatchlist({
   query: FEATURED_WATCHLIST_QUERY,
