+++ src/ducks/Watchlists/GetWatchlists.js
@@ -1,5 +1,5 @@
 import getEnhancedWatchlist from './getEnhancedWatchlist'
-import { ALL_WATCHLISTS_QUERY } from './../../components/WatchlistPopup/WatchlistGQL'
+import { ALL_WATCHLISTS_QUERY } from '../../queries/WatchlistGQL'
 
 export default getEnhancedWatchlist({
   query: ALL_WATCHLISTS_QUERY,
