+++ stories/NavbarStory.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import StoryRouter from 'storybook-react-router'
 import { Provider } from 'react-redux'
 import { MockedProvider } from 'react-apollo/test-utils'
-import { ALL_WATCHLISTS_QUERY } from './../src/components/WatchlistPopup/WatchlistGQL'
+import { ALL_WATCHLISTS_QUERY } from '../src/queries/WatchlistGQL'
 import { Panel } from '@santiment-network/ui'
 import Navbar from './../src/components/Navbar/Navbar'
 import { NavbarProfileDropdown } from './../src/components/Navbar/NavbarProfileDropdown'
